+++ CHANGELOG.md
@@ -6,14 +6,17 @@
 - Placeholder completions
   - Placeholder Support for nodes, completions and hovers
 - Hover and Go To Definition support for Placeholder Segments that use a Mob/Item/Custom Placeholder/Stat Id (for instance, <caster.stat.example> or <placeholder.test>)
-- Mob Options Completions
+- RandomSpawns File Type Recognition, with Nodes, Completions, Hovers and all the good stuff
+- Mob Options, DropOptions and DisplayOptions Completions
+  
+#### Commands
 - `MythicScribe.putSelectionInsideInlineMetaskill` command and context menu item to put the selected text inside of an inline metaskill
-- Adding a Custom Enum Dataset with the same name as another loaded Enum Dataset will now merge them
 - `MythicScribe.colorProviderOptions` config to customize the behavior of the color provider
 
 ### Changed
 - Displayed information regarding Formatter errors should now be more concise and meaningful
 - File Completions now include completions for certain keys with arbitrary values (For example, the Furniture's States or a Mob's Trades)
+- Adding a Custom Enum Dataset with the same name as another loaded Enum Dataset will now merge them
 
 ### Fixed
 - Condition Actions having no syntax highlighting